Add GetIntModByString#2

+++ vars/GetIntByString.groovy
@@ -2,5 +2,5 @@ import java.math.BigInteger;
 
 def call (String text, Int mod){
     BigInteger bigInt = new BigInteger(text.getBytes());
-    return mod(bigInt);
+    return mod(bigInt).intValue();
 }
